chore: Rename logUtils log-utils; deleted loggingUtils
