+++ ingress/src/routes/logs.ts
@@ -9,8 +9,7 @@ import {
   setFilterCursor,
   setFilterName,
   setLogLinks,
-  DEFAULT_LIMIT,
-} from "../utils/logUtils";
+} from "../utils/log-utils";
 import { isValidDeadLetterType } from "../utils/utils";
 import { QueryFilter, HateoasLogCollection } from "types/types";
 
@@ -55,7 +54,7 @@ router.get("/events", async (req: Request, res) => {
 
   try {
     setFilterTimestamp(req, filter);
-    setFilterName(req, filter, '/events')
+    setFilterName(req, filter, "/events");
   } catch (e) {
     if (!(e instanceof Error)) return;
 
@@ -94,7 +93,7 @@ router.get("/functions", async (req: Request, res) => {
 
   try {
     setFilterTimestamp(req, filter);
-    setFilterName(req, filter, '/functions');
+    setFilterName(req, filter, "/functions");
   } catch (e) {
     if (!(e instanceof Error)) return;
